+++ Tunny/WPF/Common/SelectSamplerType.cs
@@ -16,5 +16,6 @@
         QMC,
         BruteForce,
         DE,
+        cTPE,
     }
 }